+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add the basic TypeScript definition ([#40](https://github.com/marp-team/marpit/pull/40))
+- Support heading divider ([#41](https://github.com/marp-team/marpit/pull/41))
 
 ## v0.0.8 - 2018-06-28